Hide measure label when behind camera (#653)
Don't render the measure label when it is behind the camera.
Previously it was being projected back on to the screen even though it was behind the camera.
Now does a dot product to check.
Closes #647

+++ trview.app/Windows/CollapsiblePanel.cpp
@@ -107,7 +107,7 @@ namespace trview
         }
         else if (message == WM_GETMINMAXINFO)
         {
-            RECT rect{ 0, 0, static_cast<LONG>(_ui->size().width), _initial_size.height };
+            RECT rect{ 0, 0, static_cast<LONG>(_ui->size().width), static_cast<LONG>(_initial_size.height) };
             AdjustWindowRect(&rect, window_style, FALSE);
 
             MINMAXINFO* info = reinterpret_cast<MINMAXINFO*>(lParam);
